requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
openai
python-dotenv
requests
pydantic
numpy
pandas
tiktoken
streamlit